Experimental Results
Conducted experiments to make comparison between this version and commit
2c5f9f1394e7cabd7bcfd1186d7e63eac3ce84e5.
In this version the the tranformation from GRU outputs to target
output(one-hot vector for next character) is removed. And the network
can't achieve the same performace as before.
The thougts and conclusions on the reasons why this happens is logged
below (and in the notebook as well):
If GRU is used, then the outputs of GRU shall not be directly used as
desired output without further transformations because it cannot
represent both the vector for next charater and the vector for
transitions between time steps at the same time. (l.e. A cell accpet 2
inputs, an internal state from the previous cell and the input of this
cell (which is approximated by the internal state input already), then
the RNN cell can be treated as a normal feed forward network because the
two inputs are kind of equivalent. )